+++ packages/themes/Fundu/src/Theme/FunduTheme.js
@@ -12,6 +12,7 @@ import {
 import { TextStyling } from "./theme.Texts"
 import { ImageStyling } from "./theme.Images"
 import { ButtonStyling } from "./theme.Buttons"
+import { TableStyling } from "./theme.Tables"
 import {
   InputStyling,
   LabelStyling,
@@ -37,6 +38,7 @@ import { LayoutStyling } from "./theme.Layouts"
 import { LayoutItemStyling } from "./theme.LayoutItem"
 import { LogoStyle } from "./Logo"
 import reset from "./Reset"
+import { hideVisually } from "polished"
 
 const borderRadius = "2px"
 const headerHeight = "70px"
@@ -83,7 +85,7 @@ const color = {
   wcag: "#777",
   disabled: "#c1c1c1",
   warning: "#c54059",
-  black: "#000",
+  black: "#000000",
   subtleHover: "#F7F7FF",
   link: "#1F5D9B"
 }
@@ -122,7 +124,8 @@ export default new theming.Theme({
   targetSize,
   borderRadius,
   headerHeight,
-  fontFamily
+  fontFamily,
+  hideVisually
 })
   .add(reset)
   .add(HeadingStyling)
@@ -158,3 +161,4 @@ export default new theming.Theme({
   .add(FileInputStyles)
   .add(Slider2RailStyling)
   .add(Slider2TrackStyling)
+  .add(TableStyling)