+++ packages/apps/bank-in-a-box/CHANGELOG.md
@@ -3,6 +3,15 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="8.19.0"></a>
+# [8.19.0](https://bitbucket.org/stacc-flow/bento/compare/v8.18.0...v8.19.0) (2018-11-14)
+
+**Note:** Version bump only for package @staccx/bank-in-a-box
+
+
+
+
+
 <a name="8.18.0"></a>
 # [8.18.0](https://bitbucket.org/stacc-flow/bento/compare/v8.17.0...v8.18.0) (2018-11-13)